refactor functions for roles
