+++ src/middlewares/validateToken.middleware.js
@@ -22,19 +22,3 @@ export const authRequired = (req, res, next) => {
     return res.status(500).json({ message: err.message })
   }
 }
-
-export const adminRequired = (req, res, next) => {
-  try {
-    const { token } = req.cookies
-    const decodeToken = jwt.decode(token)
-    const role = decodeToken.role
-    if (role !== 'ADMINISTRADOR') {
-      return res.status(401).json({
-        message: 'authorization denied!'
-      })
-    }
-    next()
-  } catch (err) {
-    return res.status(500).json({ message: err.message })
-  }
-}
