+++ src/routes/users.routes.js
@@ -1,6 +1,7 @@
 import { Router } from 'express'
 import { validateSchema } from '../middlewares/validator.middleware.js'
-import { authRequired, adminRequired } from '../middlewares/validateToken.middleware.js'
+import { authRequired } from '../middlewares/validateToken.middleware.js'
+import { adminRequired } from '../middlewares/validateRole.middleware.js'
 import { newUserSchema } from '../schemas/user.schema.js'
 import { newUser, getAllUsers, getOneUser, deleteUser, updateUser } from '../controllers/user.controller.js'
 const router = Router()
